added new exception updated validations in dtos added new class for controller's validation added GlobalExceptionHandler with all exception handlers edited controller class
