+++ src/main/java/school/faang/user_service/util/mentorshipRequest/exception/UnknownRejectionReasonException.java
@@ -1,0 +1,4 @@
+package school.faang.user_service.util.mentorshipRequest.exception;
+
+public class UnknownRejectionReasonException extends RuntimeException{
+}
